refactor injectedScript; add injected dir to manage injected components; improve log_util with context options;
